Fix login history variant handling
g_variant_compare() returns 0 when the values are equal, and non-zero
otherwise, which is the opposite of what the check wants.
Also, it is valid for the login history variant to be null, when the
library is working with an older daemon. Don't emit a critical in that
case.

See: https://bugzilla.gnome.org/show_bug.cgi?id=690413

https://bugs.freedesktop.org/show_bug.cgi?id=58466