daemon: Assume GDM for autologin if we can't figure it out from systemd
Right now we figure out which display manager to configure based on
reading a symlink from systemd.

This isn't full proof though. SELinux security policies may prevent it
from being readable, the system might not be using systemd, etc.

Furthermore, in the case where it fails, we currently don't set a
a GError, which leads to a crash when trying to fetch the error
message.

This commit makes accountsservice fall back to GDM when it can't
figure out what else to do. That way we maintain better backward
compatibility.

https://gitlab.gnome.org/GNOME/gnome-control-center/-/issues/2450