daemon: Fix a potential use-after-free from a reload callback
If the reload timeout is not cancelled when the `Daemon` is destroyed,
its callback may be invoked after the `Daemon` is destroyed, resulting
in a use-after-free.

Spotted by code inspection. It’s probably not triggerable in practice
because if the `Daemon` is destroyed, the process is probably ending
right now anyway.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>