accounts-daemon.service: do not pull in nss-user-lookup.target
As per the nss-user-lookup.target documentation:

All services for which the availability of the full user/group
database is essential should be ordered after this target, but not
pull it in. All services which provide parts of the user/group
database should be ordered before this target, and pull it in.

accounts-daemon is a user, not a provider, so keep the After= and
drop the Wants=

https://www.freedesktop.org/software/systemd/man/latest/systemd.special.html#nss-user-lookup.target